// src/extra/enums.ts
export enum grenadeTypes {
    SMOKE = "SMOKE",
    HE = "HE",
    FLASH = "FLASH",
    MOLOTOV = "MOLOTOV",
}

export enum grenadeSides {
    CT = "CT",
    T = "T"
}

export enum themes {
    DARK = "dark",
    LIGHT = "light"
}

export enum maps {
    MIRAGE = "MIRAGE",
    DUST2 = "DUST2",
    INFERNO = "INFERNO",
    ANUBIS = "ANUBIS",
    TRAIN = "TRAIN",
    ANCIENT = "ANCIENT",
    // VERTIGO = "VERTIGO",
    NUKE = "NUKE",
    OVERPASS = "OVERPASS",
}